requirements.txt
absl-py==0.10.0
alabaster==0.7.12
astunparse==1.6.3
Babel==2.9.0
backcall==0.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
coloredlogs==14.0
cycler==0.10.0
decorator==4.4.2
docutils==0.16
flatbuffers==1.12
gast==0.3.3
GDAL==3.0.0
google-auth==1.21.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
humanfriendly==8.2
idna==2.10
imageio==2.5.0
imagesize==1.2.0
imgaug==0.2.9
importlib-metadata==2.0.0
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
networkx==2.5
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.44
opt-einsum==3.3.0
packaging==20.4
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.8
protobuf==3.13.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
q==2.6
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.4.1
Shapely==1.7.1
six==1.15.0
snowballstemmer==2.0.0
Sphinx==3.3.1
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0rc0
tensorflow-gpu==2.4.0
termcolor==1.1.0
tifffile==2020.9.3
tqdm==4.55.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.25.10
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
